requirements.txt
# torch==1.6.*
# torchvision==0.7.*
# gdal==2.4.*
dask[array]==2.9.*
h5py==2.10.*
matplotlib==3.1.*
numpy==1.21.*
netcdf4==1.5.*
opencv-python==4.2.*
pytorch-ignite==0.4.*
scikit-learn==0.23.*
scipy==1.5.*
tensorboardX==2.1.*
xarray==0.16.*
portalocker==2.0.*
pandas==1.1.*
denseweight==0.1.*
